refactor: comportamento ao salvar/atualizar e deletar

+++ src/main/java/com/ada/dynamo/service/ColunaService.java
@@ -2,6 +2,7 @@ package com.ada.dynamo.service;
 
 import com.ada.dynamo.dto.request.ColunaRequest;
 import com.ada.dynamo.dto.response.ColunaResponse;
+import com.ada.dynamo.exception.ItemComAssociassaoException;
 import com.ada.dynamo.exception.ItemNaoEncontradoException;
 import com.ada.dynamo.model.Coluna;
 import com.ada.dynamo.repository.ColunaRepository;
@@ -44,11 +45,11 @@ public class ColunaService implements ServiceContract<ColunaRequest, ColunaRespo
     }
 
     @Override
-    public ColunaResponse create(ColunaRequest request) {
-        quadroService.findModelById(request.getQuadroId());
+    public ColunaResponse create(ColunaRequest colunaRequest) {
+        quadroService.findModelById(colunaRequest.getQuadroId());
 
-        Coluna colunaModel = mapToModel(request);
-        colunaModel.setId(String.format("%s#%s", request.getQuadroId(), UUID.randomUUID()));
+        Coluna colunaModel = mapToModel(colunaRequest);
+        colunaModel.setId(String.format("%s#%s", colunaRequest.getQuadroId(), UUID.randomUUID()));
         Coluna coluna = repository.save(colunaModel);
 
         return mapToResponse(coluna);
@@ -56,6 +57,11 @@ public class ColunaService implements ServiceContract<ColunaRequest, ColunaRespo
 
     @Override
     public void delete(String id) {
+        List<Coluna> partial = repository.findByPartialId(id);
+        if (!partial.isEmpty()) {
+            throw new ItemComAssociassaoException(String.format("O id %s está associado a outros itens e não pode ser deletado", id));
+        }
+
         Coluna coluna = findModelById(id);
         repository.delete(coluna);
     }
@@ -63,20 +69,22 @@ public class ColunaService implements ServiceContract<ColunaRequest, ColunaRespo
     @Override
     public ColunaResponse update(ColunaRequest colunaRequest, String id) {
         findModelById(id);
+        quadroService.findModelById(colunaRequest.getQuadroId());
         Coluna colunaModel = mapToModel(colunaRequest);
         colunaModel.setId(id);
         return mapToResponse(repository.save(colunaModel));
     }
 
     private ColunaResponse mapToResponse(Coluna coluna) {
-        var colunaResponse = new ColunaResponse();
+        ColunaResponse colunaResponse = new ColunaResponse();
         BeanUtils.copyProperties(coluna, colunaResponse);
         return colunaResponse;
     }
 
     private Coluna mapToModel(ColunaRequest request) {
-        var coluna = new Coluna();
+        Coluna coluna = new Coluna();
         BeanUtils.copyProperties(request, coluna);
+        coluna.setTipo("COLUNA");
         return coluna;
     }
 }
